% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_lc.R
\name{betahatse.fast}
\alias{betahatse.fast}
\title{Calculate estimates of the covariance matrix for beta(t).}
\usage{
betahatse.fast(betahat, times, status, m, h, evalt)
}
\value{
An N x p x p matrix, where N is the number of evaluation times and
  p is the number of covariates.
}
\description{
To help with overall computational efficiency of \code{hdslc}, this function
will return multiple covariance matrices - one covariance matrix for each
requested evaluation time.
}
\details{
See Cai and Sun (2003, Scandinavian Journal of Statistics) for details on the
local constant estimator for beta(t). The user will not typically interact
with this function. The function is wrapped by \code{hdslc}, and the output
is used by \code{hdslcse.fast}.
}
\keyword{internal}

