% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cox_lc.R
\name{finda}
\alias{finda}
\title{Estimate the time-varying coefficients from a local-in-time Cox model}
\usage{
finda(tt, times, status, covars, start = rep(0, ncol(covars)), h = 400, ...)
}
\arguments{
\item{tt}{Time to estimate beta(t) at}

\item{times}{A vector of observed follow up times.}

\item{status}{A vector of status indicators, usually 0=alive, 1=dead.}

\item{covars}{A matrix or data frame of numeric  covariate values, with a
column for each covariate and each observation is on a separate row.}

\item{start}{A vector of length p of starting values to be passed to
\code{optim} for the numerical optimization procedure. p is the number of
covariates. Defaults to all zeroes.}

\item{h}{A single value on the time scale representing the bandwidth to use.}

\item{...}{Additional parameters to pass to \code{optim}.}
}
\value{
A vector of length p, where p is the number of covariates. The vector
is the estimated beta(t) from the local-in-time Cox model at time \code{tt}.
}
\description{
\code{finda} estimates the time-varying coefficients beta(t) at a single time
from a local-in-time Cox model. Think of it as a Cox model where the
the coefficients are allowed to vary with time. Further details can be found
in Cai and Sun (2003, Scandinavian Journal of Statistics) and Tian Zucker
Wei (2005, JASA).
}
\details{
The naming of the function \code{finda} stands for "find a(t)", where "a(t)"
is the notation used in Cai and Sun (2003, Scandinavian Journal of
Statistics) to represent the time-varying Cox model coefficients. I also
refer to "a(t)" as "beta(t)" through the documentation.

The user typically will not interact with this function, as \code{finda} is
wrapped by \code{hdslc}.
}

