% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hds_hat.R
\name{hds}
\alias{hds}
\title{Hazard discrimination summary estimator}
\usage{
hds(times, status, m, evaltimes = times[order(times)], se = TRUE)
}
\arguments{
\item{times}{A vector of observed follow up times.}

\item{status}{A vector of status indicators, usually 0=alive, 1=dead.}

\item{m}{A matrix or data frame of covariate values, with a column for each
covariate and each observation is on a separate row. Non-numeric values
are acceptable, as the values will be transformed into a numeric model
matrix through \code{survival::coxph}.}

\item{evaltimes}{A vector of times at which to estimate HDS. Defaults to all
the times specified by the \code{times} vector. If there are a lot of
observations, then you may want to enter in a sparser vector of times for
faster computation.}

\item{se}{TRUE or FALSE. TRUE: calculate and return standard error estimates.
FALSE: do not calculate standard errors estimates and return NAs. Defaults
to TRUE. May want to set to FALSE to save computation time if using this
function to compute bootstrap standard errors.}
}
\value{
A data frame with three columns: 1) the evaluation times, 2) the HDS
  estimates at each evaluation time, and 3) the standard error estimates at
  each evaluation time
}
\description{
\code{hds} returns HDS estimates at all specified evaluation times
}
\details{
A wrapper for \code{hds_t}. Since \code{hds_t} only estimates HDS at one time
point, this function calls \code{hds_t} multiple times to estimate the entire
HDS curve. It is the main function the user will interact with in this
package.

The covariate values \code{m} are centered for numerical stability. This is
particularly relevant for the standard error calculations.
}
\examples{
hds(times = survival::pbc[1:312, 2], status = (survival::pbc[1:312, 3]==2)*1,
    m = survival::pbc[1:312, 5])
}

