% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hds_hat.R
\name{hds_t}
\alias{hds_t}
\title{Hazard discrimination summary estimate at one time point}
\usage{
hds_t(t, L0hat, betahat, m)
}
\arguments{
\item{t}{The time at which to calculate HDS}

\item{L0hat}{A data frame with variable names of hazard and time. Typically
the object returned by \code{basehaz}.}

\item{betahat}{A vector of coefficient estimates from the Cox model.
Typically the \code{coefficients} value from the \code{coxph.object} object
returned by \code{coxph}.}

\item{m}{A numeric matrix of covariate values, with a column for each
covariate and each observation is on a separate row.}
}
\description{
\code{hds_t} estimates HDS at time t under the PH assumption
}
\details{
The user typically will not interact with this function. Rather, \code{hds}
is a wrapper for this function and is what the user typically will use.
}

