% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hds_lc_hat.R
\name{hdslc}
\alias{hdslc}
\title{Hazard discrimination summary estimator}
\usage{
hdslc(times, status, m, evaltimes = times[order(times)], h = 1.06 *
  sd(times) * (length(times)^(-0.2)), se = TRUE)
}
\arguments{
\item{times}{A vector of observed follow up times.}

\item{status}{A vector of status indicators, usually 0=alive, 1=dead.}

\item{m}{A matrix or data frame of covariate values, with a column for each
covariate and each observation is on a separate row. Non-numeric values
are acceptable, as the values will be transformed into a numeric model
matrix through \code{survival::coxph}.}

\item{evaltimes}{A vector of times at which to estimate HDS. Defaults to all
the times specified by the \code{times} vector. If there are a lot of
observations, then you may want to enter in a sparser vector of times for
faster computation.}

\item{h}{A single numeric value representing the bandwdith to use, on the
time scale. The default bandwidth is a very ad hoc estimate using
Silverman's rule of thumb}

\item{se}{TRUE or FALSE. TRUE: calculate and return standard error estimates.
FALSE: do not calculate standard errors estimates and return NAs. Defaults
to TRUE. May want to set to FALSE to save computation time if using this
function to compute bootstrap standard errors.}
}
\value{
A data frame with three columns: 1) the evaluation times, 2) the HDS
  estimates at each evaluation time, and 3) the standard error estimates at
  each evaluation time
}
\description{
\code{hdslc} returns local constant HDS estimates at all specified evaluation
times
}
\details{
A local constant version of \code{hds}. While \code{hds} estimates HDS(t)
assuming the Cox proportional hazards model, \code{hdslc} estimates HDS(t)
using a relaxed, local-in-time Cox model. Specifically, the hazard
ratios are allowed to vary with time. See Cai and Sun (2003, Scandinavian
Journal of Statistics) and Tian Zucker Wei (2005, JASA) for details on the
local-in-time Cox model.

Point estimates use \code{hdslc.fast} and standard errors use
\code{hdslcse.fast}. \code{hdslc.fast} requires an estimate of beta(t) (time-varying
hazard ratio), which is estimated using \code{finda()}; and subject specific
survival, which is estimated using sssf.fast(). \code{hdslcse.fast} the same
and in addition standard error estimates of beta(t), which are estimated
using \code{betahatse.fast()}.

The covariate values \code{m} are centered for numerical stability. This is
particularly relevant for the standard error calculations.
}
\examples{
hdslc(times = survival::pbc[1:312, 2],
      status = (survival::pbc[1:312, 3]==2)*1,
      m = survival::pbc[1:312, 5])
}

